Swap PGOP to f32

+++ spatula/boosop.py
@@ -204,8 +204,8 @@ class BOOSOP:
         )
 
         self._order, self._rotations = self._cpp.compute(
-            dist.astype(np.float64),
-            neighbors.weights.astype(np.float64),
+            dist.astype(np.float32),
+            neighbors.weights.astype(np.float32),
             np.ascontiguousarray(neighbors.neighbor_counts.astype(np.int32)),
             np.uint32(m),
             np.conj(self._ylms(l, m)),
@@ -221,9 +221,9 @@ class BOOSOP:
                 m=refine_m, weights=True, cartesian=True
             )
             self._order = self._cpp.refine(
-                dist.astype(np.float64),
+                dist.astype(np.float32),
                 self._rotations,
-                neighbors.weights.astype(np.float64),
+                neighbors.weights.astype(np.float32),
                 np.ascontiguousarray(neighbors.neighbor_counts.astype(np.int32)),
                 np.uint32(refine_m),
                 np.conj(self._ylms(refine_l, refine_m)),
