+++ src/util/QlmEval.h
@@ -29,8 +29,8 @@ class QlmEval {
      * ordering of the first dimension is in accending order of \f$ l \f$ and \f$ m \f$.
      */
     QlmEval(unsigned int m,
-            const double* positions,
-            const double* weights,
+            const float* positions,
+            const float* weights,
             const std::complex<double>* ylms,
             size_t n_quad_points,
             size_t n_lms);
